requirements.txt
scipy
numpy<2
matplotlib
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tqdm
opencv-python==4.10.0.84
pillow==10.4.0
h5py
